Homework9
Step 1. Connect Room to project
Step 2. Create DAOs for Movie objects
Step 3. When a screen opens, request the data from db first, send it to UI and only then request data from network and send it to the UI
Step 4. After loading the movies from the API - save them into the db, so they can be read in the future screen openings
Steps 5-6. Repeat steps 3-4 for the Movie Details screen